Adding mapping guidelines
Motivation

To extend the possibilities to apply PaNET, mappings to relevant established sources are highly useful. In order to provide a structured approach to create mappings, providing some guidelines would be useful.
Modification

An md file with guidelines has been added, as well as a template in csv format.
Result

Future mappings will be easier to create and will have a well-defined structure.
Related Issues

closes #288